+++ pkgs/watcher/test/utils.dart
@@ -63,8 +63,13 @@ Future<void> startWatcher({String? path}) async {
     final normalized = p.normalize(p.relative(path, from: d.sandbox));
 
     // Make sure we got a path in the sandbox.
-    assert(p.isRelative(normalized) && !normalized.startsWith('..'),
-        'Path is not in the sandbox: $path not in ${d.sandbox}');
+    if (!p.isRelative(normalized) || normalized.startsWith('..')) {
+      // The polling watcher can poll during test teardown, signal using an
+      // exception that it will ignore.
+      throw FileSystemException(
+        'Path is not in the sandbox: $path not in ${d.sandbox}',
+      );
+    }
 
     var mtime = _mockFileModificationTimes[normalized];
     return mtime != null ? DateTime.fromMillisecondsSinceEpoch(mtime) : null;
@@ -174,6 +179,23 @@ Matcher isModifyEvent(String path) => isWatchEvent(ChangeType.MODIFY, path);
 /// [path].
 Matcher isRemoveEvent(String path) => isWatchEvent(ChangeType.REMOVE, path);
 
+/// Takes the first event omitted during [duration], or returns `null` if there
+/// is none.
+Future<WatchEvent?> waitForEvent({
+  Duration duration = const Duration(seconds: 1),
+}) async {
+  final result = await _watcherEvents.peek
+      .then<WatchEvent?>((e) => e)
+      .timeout(duration, onTimeout: () => null);
+  if (result != null) _watcherEvents.take(1).ignore();
+  return result;
+}
+
+/// Expects that no events are omitted for [duration].
+Future expectNoEvents({Duration duration = const Duration(seconds: 1)}) async {
+  expect(await waitForEvent(duration: duration), isNull);
+}
+
 /// Expects that the next event emitted will be for an add event for [path].
 Future expectAddEvent(String path) =>
     _expectOrCollect(isWatchEvent(ChangeType.ADD, path));
@@ -222,6 +244,34 @@ void writeFile(String path, {String? contents, bool? updateModified}) {
     path = p.normalize(path);
 
     _mockFileModificationTimes[path] = _nextTimestamp++;
+  }
+}
+
+/// Schedules writing a file in the sandbox at [link] pointing to [target].
+///
+/// If [updateModified] is `false`, the mock file modification time is not
+/// changed.
+void writeLink({
+  required String link,
+  required String target,
+  bool? updateModified,
+}) {
+  updateModified ??= true;
+
+  var fullPath = p.join(d.sandbox, link);
+
+  // Create any needed subdirectories.
+  var dir = Directory(p.dirname(fullPath));
+  if (!dir.existsSync()) {
+    dir.createSync(recursive: true);
+  }
+
+  Link(fullPath).createSync(target);
+
+  if (updateModified) {
+    link = p.normalize(link);
+
+    _mockFileModificationTimes[link] = _nextTimestamp++;
   }
 }